reorganize functions and rename
