+++ src/arviz_base/__init__.py
@@ -7,10 +7,11 @@ _log = logging.getLogger(__name__)
 
 from arviz_base._version import __version__
 from arviz_base.base import dict_to_dataset, generate_dims_coords, make_attrs, ndarray_to_dataarray
-from arviz_base.converters import *
+from arviz_base.converters import convert_to_dataset, convert_to_datatree
 from arviz_base.datasets import clear_data_home, get_data_home, list_datasets, load_arviz_data
 from arviz_base.io_cmdstanpy import from_cmdstanpy
 from arviz_base.io_dict import from_dict
 from arviz_base.io_emcee import from_emcee
 from arviz_base.rcparams import rc_context, rcParams
+from arviz_base.reorg import extract, dataset_to_dataarray, dataset_to_dataframe
 from arviz_base.sel_utils import *
